# Use rayon parallel iterators