--- a/.gitignore
+++ b/.gitignore
@@ -1,1 +1,4 @@
 /target
+perf.data
+flam
+flamegraph.svg