--- a/Cargo.toml
+++ b/Cargo.toml
@@ -7,4 +7,4 @@ edition = "2021"
 
 [dependencies]
 tiny-skia = "0.8.2"
-# num-complex = "0.4.2"
+rayon = "1.5.3"