--- a/Cargo.toml
+++ b/Cargo.toml
@@ -8,3 +8,7 @@ edition = "2021"
 [dependencies]
 tiny-skia = "0.8.2"
 rayon = "1.5.3"
+
+# For flamegraphs
+[profile.release]
+debug = true