Improve performance by making the computations parallel using the [`rayon`](https://crates.io/crates/rayon) crate.

**Before**
```
Benchmark 1: ./target/release/mandelbrot_rs
  Time (mean ± σ):      1.363 s ±  0.025 s    [User: 1.351 s, System: 0.006 s]
  Range (min … max):    1.338 s …  1.431 s    10 runs
```

**After**
```
Benchmark 1: ./target/release/mandelbrot_rs
  Time (mean ± σ):     649.1 ms ±  12.1 ms    [User: 1931.2 ms, System: 33.6 ms]
  Range (min … max):   631.0 ms … 670.2 ms    10 runs
```